chapter6_self/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
attrs==25.4.0
aws-requests-auth==0.4.3
beautifulsoup4==4.14.2
bedrock-agentcore==0.1.2
boto3==1.40.72
botocore==1.40.72
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
cryptography==46.0.3
dill==0.4.0
docstring-parser==0.17.0
frozenlist==1.8.0
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.3
idna==3.11
importlib-metadata==8.7.0
jmespath==1.0.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
lxml==6.0.2
markdown-it-py==4.0.0
markdownify==1.2.0
mcp==1.21.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.7.0
opentelemetry-api==1.38.0
opentelemetry-instrumentation==0.59b0
opentelemetry-instrumentation-threading==0.59b0
opentelemetry-sdk==1.38.0
opentelemetry-semantic-conventions==0.59b0
packaging==25.0
pillow==11.3.0
prompt-toolkit==3.0.52
propcache==0.4.1
pycparser==2.23
pydantic==2.12.4
pydantic-core==2.41.5
pydantic-settings==2.12.0
pygments==2.19.2
pyjwt==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
python-multipart==0.0.20
readabilipy==0.3.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.28.0
s3transfer==0.14.0
six==1.17.0
slack-bolt==1.26.0
slack-sdk==3.37.0
sniffio==1.3.1
soupsieve==2.8
sse-starlette==3.0.3
starlette==0.50.0
strands-agents==1.6.0
strands-agents-tools==0.2.5
sympy==1.14.0
tenacity==9.1.2
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.5.0
uv==0.8.13
uvicorn==0.38.0
watchdog==6.0.0
wcwidth==0.2.14
webencodings==0.5.1
wrapt==1.17.3
yarl==1.22.0
zipp==3.23.0
